Names for Gargoyles, DLC NPCs & Furnace Golems, and Misc
New convention for the Valiant Gargoyles where 'Valiant' is removed on all but the Boss' Params. Also includes a ton of their SpEffects named courtesy of a lot of digging into the AI LUA and ANIBND for them.

Heavy use of matt's Entity dump to name all NPCs properly (as far as I can ensure). Includes location tags for them, the Furnace Golems, and Gargoyles.