Fix ParamCategories.cs still using old filenames
it was still using "Categories.json" for writing and for reset to default (but not reading) even though the file has been renamed to "Param Categories.json"

Additionally reworked the error checking a little bit to prevent it from silently failing if it can't find the default params file or if somehow the directory its looking for doesn't exist (should be impossible since it creates it right before, and might be overkill, but considering that it checks if its there anyway, I think its valid to at least have an error log in case it does fail.
Also renamed AC4, ACFA, ACV, & ACVD "Categories.json" to "Param Categories.json". These got missed when the rest were renamed